pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "trustlayer"
version = "0.1.0"
description = "The universal AI trust layer"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "python-multipart>=0.0.9",
    "pdfminer.six>=20221105",
]

[project.scripts]
trustlayer = "trustlayer_cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "cli*"]
